+++ CST465/Lab5/SQL/Stored Procs/StringAttribute_InsertUpdate.sql
@@ -1,7 +1,7 @@
 ﻿CREATE PROCEDURE [aaronlindberg].StringAttribute_InsertUpdate
 (
-	@AttributeId int,
-	@EventFk int,
+	@AttributeId bigint,
+	@EventFk bigint,
 	@AttributeName varchar(64),
 	@AttributeValue varchar(4096)
 )
