+++ CST465/Lab5/SQL/Tables/DateTimeAttribute.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[DateTimeAttribute] (
-    [DateTimeId]            INT          IDENTITY (1, 1) NOT NULL,
-    [EventMemoryFk]         INT          NOT NULL,
+    [DateTimeId]            BIGINT       IDENTITY (1, 1) NOT NULL,
+    [EventMemoryFk]         BIGINT       NOT NULL,
     [DateTimeAttributeName] VARCHAR (64) NOT NULL,
     [DateTimeValue]         DATETIME     NOT NULL,
     PRIMARY KEY CLUSTERED ([DateTimeId] ASC),
