+++ CST465/Lab5/SQL/Tables/Property.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Property] (
-    [Id]      INT              IDENTITY (1, 1) NOT NULL,
+    [Id]      BIGINT           IDENTITY (1, 1) NOT NULL,
     [Name]    VARCHAR (64)     NOT NULL,
     [Creator] UNIQUEIDENTIFIER NULL,
     PRIMARY KEY CLUSTERED ([Id] ASC),
